lasarQuery and other updates
Updated lasarQuery function to query based on stations rather than HUCs
due to a number of missing stations in the XLU_AREA table within LASAR2

Added a csv of all the stations in LASAR as of 2016-12-13 and the
corresponding HUCs.

Added Station Description to station grouping and summary.

Fixed a bug when bringing in spawning dates.